fix: infinite render
Initially, when you start the dev server locally, we'd get an infinite render, forcing the app to not render anything at all on the scree, but a "loading..." text.

This is because `BskyAuthProvider` and `W3UIProvider` were created(dynamically imported) in `RootProviders`. Moving them out of the component's render cycle fixed this problem